server and app file are now in 1 file called index (didnot dlete the original file content just added code for routes and DB connection and global err handling
