+++ backend/src/index.js
@@ -1,0 +1,37 @@
+const express = require("express");
+const cors = require("cors");
+const dotenv = require("dotenv");
+dotenv.config();
+const authRoutes = require("./routes/auth.routes");
+const connectDB = require("./config/db.js");
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+connectDB();
+
+app.use("/api/v1/auth", authRoutes);
+app.get("/api", (req, res) => {
+  res.send("<h1>Hello From Egypt!</h1>");
+});
+
+// Handle unmatched routes
+app.use((req, res) => {
+  res.status(404).json({
+    status: "ERROR",
+    data: null,
+    message: "Route not found",
+  });
+});
+
+// Global error handler
+app.use((err, req, res, next) => {
+  console.error("Global Error Handler:", err.message);
+  res.status(err.status || 500).json({ error: err.message });
+});
+
+const PORT = process.env.PORT || 3000;
+app.listen(PORT, () => {
+  console.log(`Express server running at http://localhost:${PORT}/`);
+});
