--- backend/src/server.js
@@ -1,15 +1,0 @@
-require("dotenv").config();
-const http = require("http");
-const app = require("./app");
-const connectDB = require("./config/db");
-
-const PORT = process.env.PORT || 5000;
-
-// Connect to MongoDB
-connectDB();
-
-const server = http.createServer(app);
-
-server.listen(PORT, () => {
-  console.log(`🚀 Server running on port http://localhost:${PORT}`);
-});
